Bugfix: SPI ist manchmal in Overflow gekommen, weil Zeitfenster um RXreg auszulesen überschritten wurde durch Timer-Tick-Interrupt. In der Tick-Funktion wurde ein Pin getoggelt, was Zeit kostet und dann random SPI zu lange blockiert hat, sodass RXreg nicht zügig genug ausgelesen wurde befor neue Daten im Shift-Reg lagen. Fix: es hätte gereicht das Pin-Toggeln auszumachen, aber saubere Lösung war die Interrupt-Prio so zu ändern, das SPI höhere Prio hat als Timer.

+++ Core/Src/scheduler.c
@@ -5,11 +5,11 @@
 #include "stm32f1xx_hal_gpio.h"
 
 volatile Task task_list[MAX_TASKS];
-volatile uint16_t current_time_us = 0;
+volatile uint32_t current_time_us = 0;
 
 
 //changes next_execution-time and sets lies_in_future-flag. No new task is added.
-void scheduler_calc_next_execution(Task* task, uint16_t delay_us){
+void scheduler_calc_next_execution(Task* task, uint32_t delay_us){
 	task->next_execution = current_time_us + delay_us;
 	task->lies_in_future_time_frame = task->next_execution < current_time_us;
 }
@@ -22,7 +22,7 @@ void scheduler_init(void) {
     }
 }
 
-void scheduler_add(task_func_t func, uint16_t delay_us, uint8_t run_count) {
+void scheduler_add(task_func_t func, uint32_t delay_us, uint8_t run_count) {
     for (uint8_t i = 0; i < MAX_TASKS; i++) {
         if (!task_list[i].active) {
             task_list[i].function = func;
@@ -46,7 +46,7 @@ void scheduler_remove(task_func_t func) {
 
 void scheduler_run(void) {
     for (uint8_t i = 0; i < MAX_TASKS; i++) {
-        if (task_list[i].active && (int16_t)(current_time_us - task_list[i].next_execution) >= 0 && !task_list[i].lies_in_future_time_frame) {
+        if (task_list[i].active && (int32_t)(current_time_us - task_list[i].next_execution) >= 0 && !task_list[i].lies_in_future_time_frame) {
             task_list[i].function();
             if (task_list[i].remaining_runs > 1 || task_list[i].remaining_runs == 255) {
                 if (task_list[i].remaining_runs != 255) {
@@ -63,7 +63,7 @@ void scheduler_run(void) {
 // ISR called every 5 microseconds
 void scheduler_tick(void) {
 	HAL_GPIO_TogglePin(GPIOB, GPIO_PIN_15);
-	uint16_t prev_time = current_time_us;
+	uint32_t prev_time = current_time_us;
     current_time_us += TIMER_TICK_US;
     if(current_time_us < prev_time){
     	for(uint8_t i = 0; i < MAX_TASKS; i++){
